Add new values to the tokenset
